requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
enum-compat==0.0.3
Flask==3.0.1
future==0.18.3
idna==3.6
intuit-oauth==1.2.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
oauthlib==3.2.2
pyasn1==0.5.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-quickbooks==0.9.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
